requirements.txt
streamlit
yfinance
plotly
pandas
numpy
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-core
sentence-transformers
faiss-cpu
groq
openai
python-dotenv
